+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## Unreleased
+
+### Added
+
+* [HexaPDF::Document::Metadata] for working with metadata (reading the info
+  dictionary and writing it as well as the XMP metadata stream)
+
+
 ## 0.36.0 - 2024-01-20
 
 ### Added